Move general info tab from "Profile settings" to settings
