+++ app/src/main/java/com/example/rum8/adapters/ProfileSettingsViewPagerAdapter.java
@@ -5,19 +5,16 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
-import com.example.rum8.fragments.ProfileSettingsGeneralInfoFragment;
 import com.example.rum8.fragments.ProfileSettingsPersonalityLogisticsFragment;
 import com.example.rum8.fragments.ProfileSettingsRoommatePreferencesFragment;
 
 public class ProfileSettingsViewPagerAdapter extends FragmentPagerAdapter {
 
     private final Fragment[] fragments = {
-            new ProfileSettingsGeneralInfoFragment(),
             new ProfileSettingsPersonalityLogisticsFragment(),
             new ProfileSettingsRoommatePreferencesFragment()
     };
     private final String[] fragmentTitles = {
-            "General Info",
             "Personality Logistics",
             "Roommate Preferences"
     };
