+++ app/src/main/java/com/example/rum8/adapters/SettingsViewPagerAdapter.java
@@ -7,16 +7,19 @@ import androidx.fragment.app.FragmentPagerAdapter;
 
 import com.example.rum8.fragments.SettingsAboutMeFragment;
 import com.example.rum8.fragments.SettingsContactFragment;
+import com.example.rum8.fragments.SettingsGeneralFragment;
 import com.example.rum8.fragments.SettingsHousingFragment;
 
 public class SettingsViewPagerAdapter extends FragmentPagerAdapter {
 
     private final Fragment[] fragments = {
+            new SettingsGeneralFragment(),
             new SettingsAboutMeFragment(),
             new SettingsHousingFragment(),
             new SettingsContactFragment()
     };
     private final String[] fragmentTitles = {
+            "General",
             "About me",
             "Housing",
             "Contact"
