+++ app/src/main/java/com/example/rum8/fragments/ProfileSettingsPersonalityLogisticsFragment.java
@@ -1,6 +1,5 @@
 package com.example.rum8.fragments;
 
-import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -178,27 +177,6 @@ public class ProfileSettingsPersonalityLogisticsFragment extends Fragment implem
     }
 
     @Override
-    public void showUploadImageProgress() {
-    }
-
-    @Override
-    public void hideUploadImageProgress() {
-    }
-
-    @Override
-    public void updateUploadImagePercentage(double percengate) {
-    }
-
-    @Override
-    public void chooseImage() {
-    }
-
-    @Override
-    public void setUserProfileImage(Bitmap bitmap) {
-
-    }
-
-    @Override
     public void goToMainPage() {
     }
 
@@ -286,11 +264,5 @@ public class ProfileSettingsPersonalityLogisticsFragment extends Fragment implem
         radioGroupPersonalQuestionSeven.jumpDrawablesToCurrentState();
         radioGroupPersonalQuestionEight.jumpDrawablesToCurrentState();
     }
-
-    @Override
-    public void showDefaultImage() {
-
-    }
-
 
 }
